Renames remaining references to history
[ui tests]
