+++ frontend/app/src/components/history/LocationDisplay.vue
@@ -4,9 +4,9 @@
 
 <script lang="ts">
 import { Component, Prop, Vue } from 'vue-property-decorator';
-import { tradeLocations } from '@/components/trades/consts';
-import LocationIcon from '@/components/trades/LocationIcon.vue';
-import { TradeLocationData } from '@/components/trades/type';
+import { tradeLocations } from '@/components/history/consts';
+import LocationIcon from '@/components/history/LocationIcon.vue';
+import { TradeLocationData } from '@/components/history/type';
 import { TradeLocation } from '@/services/history/types';
 import { assert } from '@/utils/assertions';
