+++ frontend/app/src/components/history/OpenTrades.vue
